Python package added + refactoring.
