+++ gold_standard/spacy.py
@@ -1,0 +1,32 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8, vim: expandtab:ts=4 -*-
+
+"""spaCy-related stuff."""
+
+from spacy.tokens import Doc
+
+
+class WhitespaceTokenizer:
+    """
+    A tokenizer that splits on whitespaces. Required if we want to run spaCy
+    on already tokenized data.
+    """
+    def __init__(self, vocab):
+        self.vocab = vocab
+
+    def __call__(self, text):
+        words = text.split(" ")
+        spaces = [True] * len(words)
+        # Avoid zero-length tokens
+        for i, word in enumerate(words):
+            if word == "":
+                words[i] = " "
+                spaces[i] = False
+        # Remove the final trailing space
+        if words[-1] == " ":
+            words = words[0:-1]
+            spaces = spaces[0:-1]
+        else:
+            spaces[-1] = False
+
+        return Doc(self.vocab, words=words, spaces=spaces)
